refactor - change dark mode use addons
Previously, our method for implementing dark mode relied on a addon that only worked with Storybook 8. I changed our method to use Storybook's theming addon, but found it ineffective. With this change, it uses a fork of the previous dark mode addon that works with Storybook 9, since the initial addon is not maintained